+++ lizard_damage/management/commands/clean_up.py
@@ -24,26 +24,16 @@ class Command(BaseCommand):
     def handle(self, *args, **options):
         logger.info("Cleaning up scenarios which are expired...")
         now = datetime.datetime.now()
-        for damage_scenario in models.DamageScenario.objects.filter(
-                expiration_date__lte=now):
+        exp = models.DamageScenario.objects.filter(expiration_date__lte=now)
+        for damage_scenario in exp:
 
             logger.info(
                 "Deleting scenario %d (%s), tasks, events and results..." % (
                     damage_scenario.id, str(damage_scenario)))
 
+            damage_scenario.delete()
+
             SecuredPeriodicTask.objects.filter(
                 name__contains='Scenario (%05d)' % damage_scenario.id).delete()
-
-            for damage_event in damage_scenario.damageevent_set.all():
-                for damage_event_result in (
-                        damage_event.damageeventresult_set.all()):
-                    if damage_event_result.image:
-                        damage_event_result.image.delete()
-                    damage_event_result.delete()
-                if damage_event.result:
-                    damage_event.result.delete()
-
-                damage_event.delete()
-            damage_scenario.delete()
 
         logger.info("Finished.")