+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog of lizard-damage
 2.0.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Update cleanup command.
 
 
 2.0.1 (2015-03-26)